third_party/gvisor: remove old cgroups import
We don't want to use the old cgroups/v1 import, remove it.
Most code using it is no longer in use as it has been replaced by
containerd shim code, so just delete that.
Adjust the rest to work with cgroups/v3.

As K8s 1.32 drops its dependency on cgroups/v1 Gazelle mangles the deps
if we leave it in.

Change-Id: Ic95fce7a17acbddf70a1b5b2a0978b9246a17c58
Reviewed-on: https://review.monogon.dev/c/monogon/+/3720
Tested-by: Jenkins CI
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>