m/n/c/curator: fix panic caused by race
The starting/stopping of the respective follower/leader implementations
is not synchronized with the state event source. Thus it can happen that
requests get routed to an improper implementation. The follower
correctly rejected requests when the current state indicated that it was
a leader, but panic'ed if it was neither which is a valid state.

Reject requests when it is not a follower, not just if it is a leader.

Change-Id: I5cee85a44cb8a1ce1f4f28c49930abd1d14ac4e1
Reviewed-on: https://review.monogon.dev/c/monogon/+/3782
Reviewed-by: Jan Schär <jan@monogon.tech>
Tested-by: Jenkins CI