m/n/kubernetes: fix storage provisioner rate limiting
Forget() resets the rate limiter, so it should only be called when
processing has suceeded.

For example, provisioning can fail for a block volume if there is not
enough disk space for the requested size. Previously, this caused the
log to be quickly  spammed with "Failed processing item" messages, all
with "numrequeues: 0". With the fix, the retries are properly backed
off, with the requeue counter incrementing.

Change-Id: I8a31fa03fadb202205967e045d4e30f04567d9d1
Reviewed-on: https://review.monogon.dev/c/monogon/+/4007
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>