m/n/c/l/crypt: open blockdevs read-only for discovery
Use the new blockdev capabilities to only open the block devices in
read-only mode for partition discovery. This allows us to disable writes
to mounted partitions in Linux 6.12 and not get a spurious warning for
every boot. It's also generally good practice as we don't want to write
at that stage anyways.

Change-Id: If8dd9b49ae593aac6f0a25d439baa0b7d60d7ffe
Reviewed-on: https://review.monogon.dev/c/monogon/+/3986
Reviewed-by: Jan Schär <jan@monogon.tech>
Tested-by: Jenkins CI