m/cli/takeover/e2e: test with metroctl
This extends the starting end of the takeover end-to-end test by
executing metroctl instead of duplicating some of the logic in the test.

VM logs are now printed with %q such that that terminal escape codes are
printed escaped instead of raw. Without this, parts of the logs are
obscured after a new kernel boots.

Change-Id: I05de3eb2ce142a99815eba6b978dad92772fe10f
Reviewed-on: https://review.monogon.dev/c/monogon/+/4034
Tested-by: Jenkins CI
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>