osbase/net/sshtakeover: fix upload cancelation
This updates sftp to the latest version, which has various improvements.

The previous method of canceling uploads doesn't work anymore, because
Close tries to take a lock already held by ReadFromWithConcurrency.
Even with the previous version of sftp, this wasn't a correct way to
cancel an upload, because it just sent a message to the server to close
the handle, while continuing the upload in the other goroutine.

There is a work in progress v2 of sftp where Close actually can be used
to cancel an upload: https://github.com/pkg/sftp/issues/603

For now, I think the best way to cancel an upload is to make the reader
return an error.

Change-Id: Id70ea7d35526a2a98295168de26469e28fd127dd
Reviewed-on: https://review.monogon.dev/c/monogon/+/4048
Tested-by: Jenkins CI
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>