metropolis/handbook: update for new OS image format
This updates the instructions for setting up a demo cluster. We don't
have recent release build, so I changed the instructions to building it
yourself for now. This way, we don't need to provide the image as an
argument to install, because it will be taken from runfiles. The
previous instructions didn't work anyway because the --installer flag
was missing, which is now also provided by runfiles.

Change-Id: Ie98cd5a1ab097623b4258ddfaedd2a64bcaefb67
Reviewed-on: https://review.monogon.dev/c/monogon/+/4124
Tested-by: Jenkins CI
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>