build: add support for filtered stamp file
The stable status file changes whenever a commit is made, which triggers
unnecessary rebuilds of targets which are only stamped with variables
which change less often. My idea for fixing this is to create a filtered
stamp file with only the variables which change less often, and then
make targets depend on this more stable stamp file instead of the stable
status file.

This change adds a Bazel rule for generating such a filtered stamp file,
and a patch for rules_go which allows using the filtered file in x_defs.

Change-Id: I1a98babeb0cc5edeac6a90c655117305c499f744
Reviewed-on: https://review.monogon.dev/c/monogon/+/4166
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>
Tested-by: Jenkins CI