build: switch to build settings for static build
This switches the way to select static linking for C builds from a
constraint_setting to a build_setting. The benefit of this is that the
transition to static linking no longer depends on the architecture,
because this setting is not part of the platform anymore.

The alternative would be to select the matching static platform based on
the architecture in the transition implementation. However, I did not
find a good way to obtain the architecture in a transition
implementation. You can access the platform through the settings
parameter, but it's not clear how to extract the architecture from that.
Something that would work is to define an attribute for the architecture
on the rule where the transition is used, and fill this attribute using
a select() at the point where the rule is instantiated. But this
requires changes to all rules which use the transition, which is ugly.

Change-Id: I8962e24b561b8d1724bac785cb3739cdc434d47e
Reviewed-on: https://review.monogon.dev/c/monogon/+/4172
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>