m/cli/metroctl: transition node platform
Add a platform transition to the node binary data dependencies of
metroctl. These binaries always need to be built for linux.
Additionally, the user may want to build them for a different
architecture than metroctl, which is possible with the new node_arch
flag. If the flag is not set, the architecture is inherited.

Change-Id: If4a15146392dfbc48c8e13db3bfb131f32176559
Reviewed-on: https://review.monogon.dev/c/monogon/+/4190
Tested-by: Jenkins CI
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>