treewide: build product info
This change adds the type definition and generator for product info,
which will be added to the OCI OS image to provide information about the
contents.

Here is an example product info:

{
	"id": "metropolis-node",
	"name": "Metropolis Node",
	"version": "0.1.0-dev1059",
	"variant": "x86_64-race",
	"commit_hash": "56248c1c1d5039bdf3c1043ade88f3f158ceb52b",
	"commit_date": "2025-05-08T18:26:46+00:00",
	"build_tree_dirty": true,
	"components": [
		{"id": "linux", "version": "6.12.15"},
		{"id": "kubernetes", "version": "1.32.0"}
	]
}

The product info has the same inputs and a similar purpose as the
os-release file, so they are both generated by the same build action.

Change-Id: I89d453f2d72ac9df49e404f46381cd594534f800
Reviewed-on: https://review.monogon.dev/c/monogon/+/4192
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>