metropolis: replace version stamp with product info
This removes the stamped metropolis version library and the associated
stampgo infrastructure, and replaces it with the product info file.

The info is now stored in a separate file in the rootfs, instead of
embedded in the core binary. This has the benefit that the core binary
no longer needs to be relinked when stamping info changes.

The version logging in core/main, and the tconsole are updated to show
some of the additional info from the product info.

Change-Id: Ic5ed0e3598e8da71b96748e8d7abfedff41acd3f
Reviewed-on: https://review.monogon.dev/c/monogon/+/4207
Tested-by: Jenkins CI
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>