.bazelrc: include platform in output directory
This change enables the --experimental_platform_in_output_dir flag,
which changes output directory names to includes the platform instead of
the legacy --cpu flag.

This fixes two problems: The first is that builds of the OS image for
two architectures use the same directory name "k8-fastbuild", so they
overwrite each other. The second problem is that when the OS image is
built as a depencency of metroctl, where metrocl is built for a
different platform, then the OS image is built in an output directory
like "k8-fastbuild-ST-12f4ce3d1c3f" and thus is not shared with the
direct build of the OS image.

After this change, the OS image is built under a path such as
"linux-x86_64-fastbuild", and this doesn't change when it is built as a
dependency of metroctl.

Change-Id: I74b91c927a633c1ce05226e035fb92cf7a8f3b04
Reviewed-on: https://review.monogon.dev/c/monogon/+/4205
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>
Tested-by: Jenkins CI