treewide: bump rules_go to v0.54.0
This includes a fix for the gopackagesdriver not finding standard
library files:
https://github.com/bazel-contrib/rules_go/commit/cf3c3af34bd869b864f5f2b98e2f41c2b220d6c9

The gopackagesdriver problem appeared with commit 0fd36f458729 which
enabled pure Go globally. rules_go runs `go list` to obtain standard
library files. When pure is enabled, `-compiled=true` is not passed to
the list command, and then the output does not include CompiledGoFiles,
only GoFiles (see `go help list`). But gopls always expects
CompiledGoFiles to be present.

Change-Id: Ide2eaae5aa45510cf83c7a8691cb0af3e112ebf9
Reviewed-on: https://review.monogon.dev/c/monogon/+/4211
Tested-by: Jenkins CI
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>