third_party/nix: symlink more files from host /etc
buildFHSEnv creates a tmpfs at /etc, mounts the host /etc at /.host-etc,
and then creates symlinks in /etc to /.host-etc for some files.
The list of files is here:
https://github.com/NixOS/nixpkgs/blob/24.11/pkgs/build-support/build-fhsenv-bubblewrap/default.nix#L57

This means that, after entering nix-shell, the /etc/bazel.bazelrc and
/etc/gitconfig files are missing. The list in buildFHSEnv is not
extensible by the user, so instead I create the additional symlinks in
wrapper.sh.

buildFHSUserEnv has been renamed to buildFHSEnv:
https://github.com/NixOS/nixpkgs/blob/7c0cdc746a0a1862133a49183f93cb80052b39e6/pkgs/top-level/aliases.nix#L376

Change-Id: I582156391e7636747fae1a20befb13aa290b80d6
Reviewed-on: https://review.monogon.dev/c/monogon/+/4212
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>
Tested-by: Jenkins CI