m/test/e2e/suites/ha: decrease disk size
Reduce the size of the data partition from around 3 GB to 512 MB for the
HA e2e test. This test has encryption with authentication enabled, which
means the entire data partition is initialized. We can speed this test
up and reduce disk wear by reducing the size.

The data partition needs to be at least 300 MB for mkfs.xfs.

Change-Id: Idd7a40c55acb6b722489194acef08138cdff8ca9
Reviewed-on: https://review.monogon.dev/c/monogon/+/4298
Tested-by: Jenkins CI
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>